requirements.txt
# DeskJarvis Python依赖
# Python 3.11+

# AI模型API
anthropic>=0.34.0  # Claude
openai>=1.0.0      # OpenAI/ChatGPT/DeepSeek/Grok

# 浏览器自动化
playwright>=1.48.0

# OCR识别
ddddocr>=1.4.11  # 验证码识别（必需，自动安装）
pytesseract>=0.3.10  # 通用OCR（可选增强，需要系统安装Tesseract）
Pillow>=10.0.0  # 图片处理（Tesseract OCR需要，也是其他功能的依赖）

# 工具库
python-dotenv>=1.0.0
python-docx>=1.1.0  # Word文档处理

# 记忆系统依赖
chromadb>=0.4.0              # 向量数据库（本地存储）
sentence-transformers>=2.2.0  # 文本嵌入模型（语义搜索）

# 多代理协作系统（使用内置 SimpleCrew，无额外依赖）
# 注：CrewAI 因依赖冲突已移除，使用原生实现替代

# 可选依赖（增强功能）
# Pillow>=10.0.0             # 图片处理（多模态记忆）
# pyperclip>=1.8.0           # 剪贴板操作
